requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.26.0
xgboost>=2.0.0
scikit-learn>=1.3.0
plotly>=5.17.0
python-dateutil>=2.8.0
requests>=2.31.0
python-dotenv>=1.0.0
